Added applewatch tests
