+++ Movapp.xcodeproj/xcshareddata/xcschemes/MovappScreenshots.xcscheme
@@ -20,6 +20,20 @@
                ReferencedContainer = "container:Movapp.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
+         <BuildActionEntry
+            buildForTesting = "YES"
+            buildForRunning = "YES"
+            buildForProfiling = "NO"
+            buildForArchiving = "NO"
+            buildForAnalyzing = "NO">
+            <BuildableReference
+               BuildableIdentifier = "primary"
+               BlueprintIdentifier = "B4FE7DC52A125F33009AE03F"
+               BuildableName = "MovappWatchUITests.xctest"
+               BlueprintName = "MovappWatchUITests"
+               ReferencedContainer = "container:Movapp.xcodeproj">
+            </BuildableReference>
+         </BuildActionEntry>
       </BuildActionEntries>
    </BuildAction>
    <TestAction
@@ -54,6 +68,16 @@
                ReferencedContainer = "container:Movapp.xcodeproj">
             </BuildableReference>
          </TestableReference>
+         <TestableReference
+            skipped = "NO">
+            <BuildableReference
+               BuildableIdentifier = "primary"
+               BlueprintIdentifier = "B4FE7DC52A125F33009AE03F"
+               BuildableName = "MovappWatchUITests.xctest"
+               BlueprintName = "MovappWatchUITests"
+               ReferencedContainer = "container:Movapp.xcodeproj">
+            </BuildableReference>
+         </TestableReference>
       </Testables>
    </TestAction>
    <LaunchAction
@@ -83,15 +107,6 @@
       savedToolIdentifier = ""
       useCustomWorkingDirectory = "NO"
       debugDocumentVersioning = "YES">
-      <MacroExpansion>
-         <BuildableReference
-            BuildableIdentifier = "primary"
-            BlueprintIdentifier = "B08B0DE1281FBB28000EAEF1"
-            BuildableName = "MovappUITests.xctest"
-            BlueprintName = "MovappUITests"
-            ReferencedContainer = "container:Movapp.xcodeproj">
-         </BuildableReference>
-      </MacroExpansion>
    </ProfileAction>
    <AnalyzeAction
       buildConfiguration = "Debug">
