+++ Shared/Models/SetLanguage.swift
@@ -88,4 +88,17 @@ extension SetLanguage {
     static let allCases = [csUk, ukCs, plUk, ukPl, skUk, ukSk]
 }
 
-typealias SetLanguages = [SetLanguage]
+func getSetLanguage(for locale: String) -> SetLanguage {
+    let language = Languages(rawValue: locale) ?? .cs
+
+    return preferredSetLanguage(for: language)
+}
+
+private func preferredSetLanguage(for language: Languages) -> SetLanguage {
+    switch language {
+    case .cs, .pl, .sk:
+        return SetLanguage(language: Language(main: language, source: .uk), flipFromWithTo: false)
+    case .uk:
+        return .ukCs
+    }
+}
